Extract lists of items components from Meganav
