+++ modules-config.js
@@ -4,16 +4,20 @@ module.exports = [
     directory: "core",
     components: [
       "FeaturedLink",
+      "SignOutLink",
       "Meganav",
       "Logo",
       "MeganavContentPlatform",
       "MeganavContentWhyAbly",
       "MeganavContentDevelopers",
+      "MeganavContentUseCases",
       "MeganavControlMobilePanelClose",
       "MeganavControlMobilePanelOpen",
       "MeganavControl",
       "MeganavControlMobileDropdown",
-      "MeganavContentUseCases",
+      "MeganavItemsDesktop",
+      "MeganavItemsSignedIn",
+      "MeganavItemsMobile",
     ],
   },
   {
