+++ src/core/SignOutLink/component.html.erb
@@ -1,0 +1,1 @@
+<%= link_to @session_data[:logOut][:text], @session_data[:logOut][:href], method: :delete, class: @classes %>
