+++ src/core/SignOutLink/component.rb
@@ -1,0 +1,12 @@
+require "json"
+
+module AblyUi
+  module Core
+    class SignOutLink < ViewComponent::Base
+      def initialize(session_data: {}, classes: [])
+        @session_data = session_data
+        @classes = classes
+      end
+    end
+  end
+end
